Code refactor: move class to the correct path
