+++ envergo/evaluations/views.py
@@ -1,7 +1,6 @@
 from django.conf import settings
 from django.contrib.auth.mixins import LoginRequiredMixin
 from django.core.files.storage import get_storage_class
-from django.core.files.uploadedfile import UploadedFile
 from django.db import transaction
 from django.db.models.query import Prefetch
 from django.http.response import Http404, HttpResponseRedirect
@@ -10,8 +9,6 @@ from django.urls import reverse
 from django.utils.decorators import method_decorator
 from django.views.generic import DetailView, FormView, TemplateView
 from django.views.generic.edit import CreateView
-from formtools.wizard.storage.exceptions import NoFileStorageConfigured
-from formtools.wizard.storage.session import SessionStorage
 from formtools.wizard.views import NamedUrlSessionWizardView
 from ratelimit.decorators import ratelimit
 
@@ -194,74 +191,8 @@ TEMPLATES = {
 }
 
 
-class Storage(SessionStorage):
-    def set_step_files(self, step, files):
-        if files and not self.file_storage:
-            raise NoFileStorageConfigured(
-                "You need to define 'file_storage' in your "
-                "wizard view in order to handle file uploads."
-            )
-
-        if step not in self.data[self.step_files_key]:
-            self.data[self.step_files_key][step] = {}
-
-        if not files:
-            return
-
-        for field in files.keys():
-            field_files = files.getlist(field)
-            file_dicts = []
-            for field_file in field_files:
-                tmp_filename = self.file_storage.save(field_file.name, field_file)
-                file_dict = {
-                    "tmp_name": tmp_filename,
-                    "name": field_file.name,
-                    "content_type": field_file.content_type,
-                    "size": field_file.size,
-                    "charset": field_file.charset,
-                }
-                file_dicts.append(file_dict)
-
-            self.data[self.step_files_key][step][field] = file_dicts
-
-    def get_step_files(self, step):
-        wizard_files = self.data[self.step_files_key].get(step, {})
-
-        if wizard_files and not self.file_storage:
-            raise NoFileStorageConfigured(
-                "You need to define 'file_storage' in your "
-                "wizard view in order to handle file uploads."
-            )
-
-        files = {}
-        for field, field_files in wizard_files.items():
-            files[field] = []
-
-            for field_dict in field_files:
-                field_dict = field_dict.copy()
-                tmp_name = field_dict.pop("tmp_name")
-                if (step, field) not in self._files:
-                    self._files[(step, field)] = UploadedFile(
-                        file=self.file_storage.open(tmp_name), **field_dict
-                    )
-
-                files[field] = self._files[(step, field)]
-        return files or None
-
-    def reset(self):
-        # Store unused temporary file names in order to delete them
-        # at the end of the response cycle through a callback attached in
-        # `update_response`.
-        wizard_files = self.data[self.step_files_key]
-        for step_files in wizard_files.values():
-            for step_field_files in step_files.values():
-                for step_file in step_field_files:
-                    self._tmp_files.append(step_file["tmp_name"])
-        self.init_data()
-
-
 class RequestEvalWizard(NamedUrlSessionWizardView):
-    storage_name = "envergo.evaluations.views.Storage"
+    storage_name = "envergo.utils.formtools_storage.MultiFileSessionStorage"
     form_list = FORMS
     file_storage = get_storage_class(settings.UPLOAD_FILE_STORAGE)()
